--- a/addons/mail/tests/discuss/__init__.py
+++ b/addons/mail/tests/discuss/__init__.py
@@ -14,6 +14,7 @@ from . import test_discuss_reaction_controller
 from . import test_discuss_res_role
 from . import test_discuss_sub_channels
 from . import test_discuss_thread_controller
+from . import test_discuss_tools
 from . import test_message_controller
 from . import test_guest_feature
 from . import test_toggle_upload