@robodoo 2025-07-22
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/219884.png)](https://mergebot.odoo.com/odoo/odoo/pull/219884)